Add ability to view messages, found major bug
I tried to make automated tests, but it turns out that all of the tests are broken! The rake file is creating and modifying bumblebeard_challenge_development.db in postgres://localhost file instead of actually modifying the database files in our own folders. This means that all automated tests are failing, while all manual tests are working. You can log onto the site and view the three messages. But the tests are all failing, becuase they are trying to access files that have not been modified.

I have asked Bo for help, and we were unable to solve the problem. Bo suggests that we consult our cohort lead to figure out what to do. Until then, we need to think about how to handle this problem.